pyproject.toml
[project]
name = "redaction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "aiohttp>=3.12.15",
    "azure-ai-textanalytics>=5.3.0",
    "dotenv>=0.9.9",
    "pymupdf>=1.26.4",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]
